Split tests in unit vs functional tests
Unit tests test individual pieces of code in isolation, and mock
everything, or as much as possible. They are useful for complex
logic, algorithms and data structures.

Functional tests test individual features, trying to replicate
as much of a real engine as possible, and then driving the API.
