+++ test/functional/platform.js
