+++ test/functional/test-classes/test_platform.js
