+++ test/functional/test-classes/test_platform.tt
