+++ test/functional/test_cloud_sync.js
