+++ test/functional/test_permissions.js
