+++ test/functional/test_remote.js
