+++ test/manual.js
